コンポーネント分割を変更
